+++ addon/components/flexberry-textbox.js
@@ -22,15 +22,6 @@ var FlexberryTextbox = FlexberryBaseComponent.extend({
   classNames: ['flexberry-textbox', 'ui', 'input'],
 
   /**
-    Array CSS class names bindings.
-    [More info.](http://emberjs.com/api/classes/Ember.Component.html#property_classNameBindings)
-    @property classNameBindings
-    @type Array
-    @readOnly
-   */
-  classNameBindings: ['readonly:disabled'],
-
-  /**
    * Path to component's settings in application configuration (JSON from ./config/environment.js).
    *
    * @property appConfigSettingsPath