+++ addon/components/object-list-view.js
@@ -18,6 +18,16 @@ export default FlexberryBaseComponent.extend(
   FlexberryLookupCompatibleComponentMixin,
   FlexberryFileCompatibleComponentMixin,
   ErrorableMixin, {
+  /**
+    Projection set by property `modelProjection`.
+
+    @property _modelProjection
+    @type Object
+    @default null
+    @private
+   */
+  _modelProjection: null,
+
   actions: {
     rowClick: function(recordWithKey, e) {
       if (this.get('readonly')) {
@@ -46,7 +56,7 @@ export default FlexberryBaseComponent.extend(
         return;
       }
 
-      var action = event.ctrlKey ? 'addColumnToSorting' : 'sortByColumn';
+      var action = e.ctrlKey ? 'addColumnToSorting' : 'sortByColumn';
       this.sendAction(action, column);
     },
 
@@ -354,8 +364,17 @@ export default FlexberryBaseComponent.extend(
    * @type Boolean
    * @readonly
    */
-  showHelperColumn: Ember.computed('showAsteriskInRow', 'showCheckBoxInRow', 'showDeleteButtonInRow', function() {
-    return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow');
+  showHelperColumn: Ember.computed(
+    'showAsteriskInRow',
+    'showCheckBoxInRow',
+    'showDeleteButtonInRow',
+    'modelProjection',
+    function() {
+    if (this.get('modelProjection')) {
+      return this.get('showAsteriskInRow') || this.get('showCheckBoxInRow') || this.get('showDeleteButtonInRow');
+    } else {
+      return false;
+    }
   }),
 
   /**
@@ -435,13 +454,34 @@ export default FlexberryBaseComponent.extend(
   ),
 
   /**
-   * Model projection which should be used to display given content.
-   *
-   * @property modelProjection
-   * @type Object
-   * @default null
+    Model projection which should be used to display given content.
+    Accepts object or name projections.
+
+    @property modelProjection
+    @type Object|String
+    @default null
    */
-  modelProjection: null,
+  modelProjection: Ember.computed('_modelProjection', {
+    get(key) {
+      return this.get('_modelProjection');
+    },
+    set(key, value) {
+      if (typeof value === 'string') {
+        let modelName = this.get('modelName');
+        Ember.assert('For define projection by name, model name is required.', modelName);
+        let modelConstructor = this.get('store').modelFor(modelName);
+        Ember.assert(`Model with name '${modelName}' is not found.`, modelConstructor);
+        let projections = Ember.get(modelConstructor, 'projections');
+        Ember.assert(`Projection with name '${value}' for model with name '${modelName}' is not found.`, projections[value]);
+        value = projections[value];
+      } else if (typeof value !== 'object') {
+        throw new Error(`Property 'modelProjection' should be a string or object.`);
+      }
+
+      this.set('_modelProjection', value);
+      return value;
+    },
+  }),
 
   /**
    * Table columns related to current model projection.
@@ -493,6 +533,14 @@ export default FlexberryBaseComponent.extend(
         }
       }
     } else {
+      if (this.currentController) {
+        if (userSettings === undefined) {
+          Ember.set(this.currentController, 'userSettings', {});
+        }
+
+        Ember.set(this.currentController.userSettings, 'colsOrder', cols);
+      }
+
       ret = cols;
     }
 
@@ -809,10 +857,19 @@ export default FlexberryBaseComponent.extend(
     this.get('objectlistviewEventsService').on('olvDeleteRows', this, this._deleteRows);
     this.get('objectlistviewEventsService').on('filterByAnyMatch', this, this._filterByAnyMatch);
 
-    if (this.get('content')) {
-      this.get('content').forEach((item, index, enumerable) => {
-        this._addModel(item);
-      });
+    let content = this.get('content');
+    if (content) {
+      if (content.get('isFulfilled') === false) {
+        content.then((items) => {
+          items.forEach((item) => {
+            this._addModel(item);
+          });
+        });
+      } else {
+        content.forEach((item) => {
+          this._addModel(item);
+        });
+      }
     }
   },
 
@@ -856,6 +913,11 @@ export default FlexberryBaseComponent.extend(
     getSettingPromise.then(function(data) {
       _this._setColumnWidths(data);
     });
+
+    // TODO: resolv this problem.
+    this.$('.flexberry-dropdown:last').dropdown({
+      direction: 'upward'
+    });
   },
 
   /**
@@ -864,8 +926,11 @@ export default FlexberryBaseComponent.extend(
    *
    * @method didRender
    */
-  didRender: function() {
+  didRender() {
     this._super(...arguments);
+
+    let currentTable = this.$('table.object-list-view');
+
     if (this.get('allowColumnResize')) {
       if (this.get('useSingleColumn')) {
         Ember.Logger.error(
@@ -873,8 +938,6 @@ export default FlexberryBaseComponent.extend(
           'can\'t be enabled at the same time.');
         return;
       }
-
-      let currentTable = this.$('table.object-list-view');
 
       // The first column has semantic class "collapsing"
       // so the column has 1px width and plugin has problems.
@@ -887,6 +950,8 @@ export default FlexberryBaseComponent.extend(
       currentTable.addClass('fixed');
 
       this._reinitResizablePlugin();
+    } else {
+      currentTable.colResizable({ disable: true });
     }
   },
 
@@ -897,18 +962,18 @@ export default FlexberryBaseComponent.extend(
    * @method _reinitResizablePlugin
    * @private
    */
-  _reinitResizablePlugin: function() {
+  _reinitResizablePlugin() {
     let currentTable = this.$('table.object-list-view');
 
     // Disable plugin and then init it again.
     currentTable.colResizable({ disable: true });
 
-    let _this = this;
     currentTable.colResizable({
-      minWidth: 70,
-      onResize: function(e) {
+      minWidth: 90,
+      resizeMode:'flex',
+      onResize: (e)=> {
         // Save column width as user setting on resize.
-        _this._afterColumnResize(e);
+        this._afterColumnResize(e);
       }
     });
   },