+++ tests/dummy/app/routes/ember-flexberry-dummy-suggestion-edit.js
@@ -1,6 +1,7 @@
 import EditFormRoute from 'ember-flexberry/routes/edit-form';
+import EditFormMixin from '../mixins/edit-form-mixin';
 
-export default EditFormRoute.extend({
+export default EditFormRoute.extend(EditFormMixin, {
   /**
     Name of model projection to be used as record's properties limitation.